refactor
- clippy
- rename methods
