+++ crates/gitbutler-core/src/lib.rs
@@ -26,10 +26,10 @@ pub mod id;
 pub mod keys;
 pub mod lock;
 pub mod path;
-pub mod path_serialization;
 pub mod project_repository;
 pub mod projects;
 pub mod reader;
+pub mod serde;
 pub mod sessions;
 pub mod ssh;
 pub mod storage;
@@ -40,22 +40,3 @@ pub mod virtual_branches;
 pub mod windows;
 pub mod writer;
 pub mod zip;
-pub mod serde {
-    use crate::virtual_branches::branch::HunkHash;
-    use bstr::{BString, ByteSlice};
-    use serde::Serialize;
-
-    pub fn as_string_lossy<S>(v: &BString, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        v.to_str_lossy().serialize(s)
-    }
-
-    pub fn hash_to_hex<S>(v: &HunkHash, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        format!("{v:x}").serialize(s)
-    }
-}
