+++ crates/gitbutler-core/src/projects/project.rs
@@ -66,7 +66,7 @@ pub struct Project {
     pub title: String,
     pub description: Option<String>,
     /// The worktree directory of the project's repository.
-    #[serde(with = "crate::path_serialization")]
+    #[serde(with = "crate::serde::path")]
     pub path: path::PathBuf,
     #[serde(default)]
     pub preferred_key: AuthKey,
