+++ crates/gitbutler-core/src/virtual_branches/base.rs
@@ -11,9 +11,9 @@ use super::{
 use crate::{
     git::{self, diff},
     keys,
-    path_serialization::wrap_path,
     project_repository::{self, LogUntil},
     projects::FetchResult,
+    serde::path::json_escape,
     users,
     virtual_branches::branch::BranchOwnershipClaims,
 };
@@ -204,7 +204,7 @@ pub fn set_base_branch(
                 BranchOwnershipClaims::default(),
                 |mut ownership, hunk| {
                     ownership.put(
-                        &format!("{}:{}", wrap_path(&hunk.file_path), hunk.id)
+                        &format!("{}:{}", json_escape(&hunk.file_path), hunk.id)
                             .parse()
                             .unwrap(),
                     );
