+++ crates/gitbutler-core/src/virtual_branches/integration.rs
@@ -7,8 +7,8 @@ use lazy_static::lazy_static;
 use super::{errors, VirtualBranchesHandle};
 use crate::{
     git,
-    path_serialization::wrap_path,
     project_repository::{self, LogUntil},
+    serde::path::json_escape,
     virtual_branches::branch::BranchCreateRequest,
 };
 
@@ -217,7 +217,7 @@ pub fn update_gitbutler_integration_with_commit(
         }
         for file in &branch.ownership.claims {
             message.push_str("   - ");
-            message.push_str(&wrap_path(&file.file_path));
+            message.push_str(&json_escape(&file.file_path));
             message.push('\n');
         }
     }
